double

+++ src/main/java/sales/HomeSale.java
@@ -13,7 +13,7 @@ public class HomeSale {
    public String property_type;
    public String strata_lot_number;
    public String property_name;
-   public int area;
+   public double area;
    public String area_type;
    public String contract_date;
    public String settlement_date;
@@ -26,7 +26,7 @@ public class HomeSale {
 
 
    public HomeSale(int property_id, String download_date, String council_name, int purchase_price, String address, int post_code,
-    String property_type, String strata_lot_number, String property_name, int area, String area_type, String contract_date, 
+    String property_type, String strata_lot_number, String property_name, double area, String area_type, String contract_date, 
     String settlement_date, String zoning, String nature_of_property, String primary_purpose, String legal_description) {
        this.property_id = property_id;
        this.download_date = download_date;
